fix indentation for bullet lists
Unlike top-level bullet lists that are not indented, nested lists
(sub-lists) need to be indented with a certain number of spaces. This
number depends on the with of the bullet mark of the parent list. For
example, if the parent list uses bullet marks of width 1 and a space,
nested bullets must have 2 spaces at the beginning of the line:

- parent item
  - child item

If the parent list uses numbered list with items of width 2, child
items must be indented with 3 spaces:

1. parent items
   - child item

Finally, this fixes one instance of a bullet item not aligned with the
rest.