escape underscore
Underscore character (_) is used in Markdown syntax to encode italic
text: _example_. When underscore is required in the text itself and is
not intended to be used as formatting, it should be "escaped" by adding
a backslash before it: \_example.

While it may work just fine in most cases, sometimes the text after
a non-escaped underscore may contain another non-escaped underscore
and the whole text segment between them will be shown as italic while
the underscores will be not shown.

Example: first let's type a word with an _underscore not intended to
mean italic text. Next let's have _another word. Now some software
may render the whole segment between the two as italic.