add metadata
It is useful to have basic metadata like primary author and published
date somewhere in the document. A nice standard for this is called
"front matter" which originated in the Jekyll static site generation
software. The format is very simple but is also very powerful by
allowing to include arbitrary data structures.