data: move files to data dir
This makes this repo consistent with the structure used in some other
repos in https://github.com/decredcommunity .

The idea is that files are grouped in "collections" (which also work as
namespaces) and each collection has its own directory in repository's
root dir.

While here, rename competitor-analysis.csv to
competitor-media-analysis.csv to reflect that the analysis is focused on
media presence.