data/monde-pr-media-coverage: sort for grouping
Sorting by first 3 columns (`date`, `outlet`, `reporter`) ascending
was an improvement that established a pretty strict order, which could
be easily restored after reorganizing the sheet.

The disadvantage of such order is it is hard to see groups of syndicated
(reposted) items. To address that, we can sort by (`date`, `reporter`)
ascending instead and exploit the observation that reporter is the same
in most syndications.

The above already yields good results in terms of grouping syndicated
items together, but it can be further improved with the new `original`
column. Sorting by `original` descending _after_ sorting by `date` and
`reporter` puts the original item at the start of each group.

Finally, to make the ordering even more stable we sort by `outlet`
ascending.

Therefore, all sorting steps are:

- sort items by date (ascending)
- sort items with equal date by reporter (ascending)
- sort items with equal (date, repoter) by original (descending)
- sort items with equal (date, reporter, original) by outlet (ascending)