data/monde-pr-media-coverage: Internet -> Article
Originally the `medium` column had "Online" for most rows, to
differentiate with "Offline" mediums like TV or Radio.

I changed "Online" to "Internet" to be more specific because terms
"online" and "offline" can mean many things.

Now with the introduction of new values like "Podcast" and "Video
Interview" we notice that _all_ of them are published in the Internet
and just stating that is not too helpful. Changing it to "Article" makes
this column more useful and consistent with the other 2 values.